:art: refactor code for checking if image cam capture and screenshot is valid

+++ CefSharpTestBrowser/WindowsFormsApp1/frmMain.cs
@@ -1695,24 +1695,14 @@ namespace WindowsFormsApp1
 
         private void retryUpload(string path)
         {
-            try
+            for (int i = 0; i < 3; i++)
             {
-                using (var bmp = new Bitmap(path))
+                if (uploadImage(path))
                 {
-                    for (int i = 0; i < 3; i++)
-                    {
-                        if (uploadImage(path))
-                        {
-                            break;
-                        }
-
-                        Thread.Sleep(30000);
-                    }
+                    break;
                 }
-            }
-            catch
-            {
-                FileUtil.deleteFile(path);
+
+                Thread.Sleep(30000);
             }
         }
 
@@ -1724,7 +1714,7 @@ namespace WindowsFormsApp1
             {
                 Monitor.Enter(path, ref lockWasTaken);
 
-                if (Globals.activity.PostScreenshot(path))
+                if (this.isImageValid(path) && Globals.activity.PostScreenshot(path))
                 {
                     FileUtil.deleteFile(path);
                     return true;
@@ -1736,6 +1726,22 @@ namespace WindowsFormsApp1
             }
 
             return false;
+        }
+
+        private bool isImageValid(string path)
+        {
+            try
+            {
+                using (var bmp = new Bitmap(path))
+                {
+                    return true;
+                }
+            }
+            catch (ArgumentException)
+            {
+                FileUtil.deleteFile(path);
+                return false;
+            }
         }
 
         private void findAndUploadFailedImages()
@@ -1822,7 +1828,7 @@ namespace WindowsFormsApp1
                     staffCam.captureImage(camFileName);
                     Thread.Sleep(1000);
                     staffCam.stopCamera();
-                    if (Globals.activity.PostCameraCapture(camFileName))
+                    if (this.isImageValid(camFileName) && Globals.activity.PostCameraCapture(camFileName))
                     {
                         FileUtil.deleteFile(camFileName);
                     }
